X13. Code simplification

+++ demetra-toolkit/demetra-toolkit-core/demetra-toolkit-basic/src/main/java/jdplus/leastsquares/QRSolution.java
@@ -39,7 +39,7 @@ public class QRSolution {
     private DoubleSeq e;
     private double ssqErr;
     
-    public double rank(){
+    public int rank(){
         return rank;
     }
